Tile coordinates

+++ src/server/object.rs
@@ -1,7 +1,7 @@
 #[derive(Debug)]
 pub struct Object {
-    pub x: f32,
-    pub y: f32,
+    pub x: u32,
+    pub y: u32,
     pub texture: String,
     pub client: Option<crate::message::Client>,
 }
